+++ crate/operator/create.py
@@ -1030,6 +1030,7 @@ def get_data_service(
                 # Default idle timeout is 60s, which kills the connection on long-running queries # noqa
                 "service.beta.kubernetes.io/aws-load-balancer-connection-idle-timeout": "3600",  # noqa
                 "service.beta.kubernetes.io/aws-load-balancer-type": "nlb",  # noqa
+                "service.beta.kubernetes.io/aws-load-balancer-cross-zone-load-balancing-enabled": "true",  # noqa
             }
         )
     elif config.CLOUD_PROVIDER == CloudProvider.AZURE: