+++ tests/test_create.py
@@ -1133,6 +1133,10 @@ class TestServiceModels:
                 "service.beta.kubernetes.io/aws-load-balancer-type"  # noqa
                 in annotation_keys
             )
+            assert (
+                "service.beta.kubernetes.io/aws-load-balancer-cross-zone-load-balancing-enabled"  # noqa
+                in annotation_keys
+            )
         if provider == "azure":
             assert (
                 "service.beta.kubernetes.io/azure-load-balancer-disable-tcp-reset"